pyproject.toml
[project]
name = "mapdisplay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-storage-blob>=12.28.0",
    "fastapi[standard]>=0.128.5",
    "iconfig-py>=0.1.10",
    "loguru>=0.7.3",
    "matplotlib>=3.8.0",
    "orjson>=3.11.7",
    "pandas>=3.0.0",
    "pillow>=10.0.0",
    "panel[fastapi]>=1.8.7",
    "pydantic>=2.12.5",
    "pydotenv>=0.0.7",
    "pytest>=9.0.2",
    "scipy>=1.10.0",
    "xarray>=2026.1.0",
    "netcdf4>=1.7.4",
    "zarr>=3.1.5",
    "cupy-cuda13x>=13.6.0",
]

[tool.setuptools]
packages = ["md", "md.preprocess", "md.storage"]

[tool.setuptools.package-dir]
"" = "src"
